Single-precision SSE `advance_angular' calculation. Decrease maximum frame time (advance_angular blows up when w*dt is too large). Fit the tank exactly to the viewing frustum but leave an air gap at the front.

+++ src/rodrigues.h
@@ -12,8 +12,8 @@
 // Function compute puts the GL modelview matrix into f.
 // Functions advance_* do inertial motion for time dt.
 
-void compute (float (& f) [16], const float (& x) [4], const double (& u) [4], float r);
+void compute (float (& f) [16], const float (& x) [4], const float (& u) [4], float r);
 void advance_linear (float (* x) [4], const float (* v) [4], unsigned count, float dt);
-void advance_angular (double (* u) [4], float (* w) [4], unsigned count, float dt);
+void advance_angular (float (* u) [4], float (* w) [4], unsigned count, float dt);
 
 #endif
